增加审核记录的存储

+++ ruoyi-system/src/main/java/com/ruoyi/system/domain/dto/LectureAuditDTO.java
@@ -4,7 +4,7 @@ import javax.validation.constraints.Size;
 
 public class LectureAuditDTO {
     @NotNull(message = "讲座记录ID不能为空")
-    private Long reportId;
+    private Integer reportId;
 
     @NotNull(message = "审核状态不能为空")
     private Integer auditStatus;
@@ -12,11 +12,11 @@ public class LectureAuditDTO {
     @Size(max = 500, message = "审核意见不能超过500个字符")
     private String auditRemark;
 
-    public @NotNull(message = "讲座记录ID不能为空") Long getReportId() {
+    public @NotNull(message = "讲座记录ID不能为空") Integer getReportId() {
         return reportId;
     }
 
-    public void setReportId(@NotNull(message = "讲座记录ID不能为空") Long reportId) {
+    public void setReportId(@NotNull(message = "讲座记录ID不能为空") Integer reportId) {
         this.reportId = reportId;
     }
